requirements.txt
.
pytest
types-requests
isort
twine
build
python-socketio
aiohttp
